Refactor: Shared module path
